requirements.txt
youtube-transcript-api
google-generativeai
dotenv